refactor(*): move loadSettings to index
