+++ src/app.js
@@ -3,7 +3,7 @@ require('./functions/db');
 const fastify = require('fastify');
 const session = require('@fastify/session');
 const MongoStore = require('connect-mongodb-session')(session);
-const { loadSettings } = require('./functions/loadSettings');
+const { loadSettings } = require('./functions');
 const { join } = require('path');
 
 const settings = loadSettings();
